EWM-283. BarcodeAddress widget
